app done

+++ app/src/main/java/ru/leushinilya/loftmoney/remote/RemoteItem.java
@@ -19,7 +19,7 @@ public class RemoteItem {
     private String name;
 
     @SerializedName("price")
-    private double price;
+    private float price;
 
     @SerializedName("type")
     private String type;
@@ -35,7 +35,7 @@ public class RemoteItem {
         return name;
     }
 
-    public double getPrice() {
+    public float getPrice() {
         return price;
     }
